Updated with some changes
New movement option: Jumping
Letters are now always valid when they spawn. (instead of random with more weight on vowels and commen letters) This means if you are able to dodge for long enough, the next letter is guarenteed to work (although it might be for a strange word you've never heard of) This greatly decreases difficulty of forming short (like 3-4 letter) words, but long words are still quite challenging because the screen can get quite busy with lots of special letters.